Substantial rework of the server.
The plan is to just keep things encoded in vcal format, rather than
translating between 5 different things. We also don't need to fully
parse the vcal data into separate tables; we can keep it together as a
blob and only pull out bits we actually want to query. This will make
things easier.

I've junked selda in favor of a simple wrapper module around
sqlite-simple.

In the short term this break the client; need to rework the encoders &
decoders to do the right thing.