Start working on a performance issue.
Right now, when collecting events to display on the main page, we have
to linearly scan each recurrence rule from the original start date of an
event up through the present. For recurring events that have been going
on for a long time, this could get expensive, and I've seen noticeable
latency when loading the front page that wasn't there before I started
implementing support for recurring events.

So the plan is to be smarter about finding our starting point. This
patch adds a helper function (commented out, to avoid failing the build
because it's unused), which we'll eventually integrate into the rest of
the code to search for that start point.