More perf improvements.
- Make Occurrences.merge O(n log n), rather than O(n^2).
- Filter out the first occurrences of events *before* merging.

The first of these didn't make a perceptible impact on perf; that's what
I get for not profiling first. But it could matter for larger calendars,
so I might as well keep it since I've done the work anyway.

The latter made a much bigger difference, as the calendar I'm working
with has many one-time-only events that are filtered out by this process
and thus don't have to be processed by the merge.