Fix the issue with a skewed monthly event.
DTSTAMP is supposed to be a modification/update date, so we should
basically never use it as the date of the event itself.

We're now a bit smarter about some cases. Some of this is still
heuristic and I don't know how well we can ultimately do given the
constraints of the iCalendar data model (or the way humans keep track of
time generally...) but this is an improvement.