Partial work on richer repeat rules.
This works on changing the data type representations in the client and
on the wire to support multiple rules and intervals other than 1.

Submitting events still doesn't do the right thing, and the UI doesn't
render differently. We'll probably need to evolve past using form data
sooner or later, maybe I should tackle that next.