Start rewriting client side OCaml in Elm
Bucklescript is unmaintained; ReScript decided it wanted to be more like
javascript so it broke a bunch of compatibility, and I don't want to
fuss with updating it, or deal with the treadmill.

There's a fork, but the fork hasn't been updated in ~6 months...

It's only ~700 lines of code, and it's using very elm like APIs so far.
Let's just re-write it in Elm.