UI: check that start time is before end time
Before creating/editing an event. If these are ordered incorrectly, the
submit button will be disabled.

Fixes #32

Note that the backend doesn't enforce this; this is a concious decision:
you would separately have to deal with bad data from iCalendar imports,
and there may already be such events in the wild.